refactor(ui): extract requests-card component with unread count badge
Extract the secret requests table, status icons, and delete dialog
into a standalone requests-card component, matching the secrets-card
pattern. Load unread response count in parallel via Promise.all.
